Moved js api calls to another file
