+++ templates/index.html
@@ -3,7 +3,7 @@
     <head>
         <meta charset="UTF-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <script src="src/script/main.js"></script>
+        <script type="module" src="src/script/main.js"></script>
         <link rel="stylesheet" href="vendor/twbs/bootstrap/dist/css/bootstrap.min.css" />
         <title>To-Do</title>
     </head>
